requirements.txt
# Core dependencies
streamlit>=1.31.0
chromadb>=0.4.22
langchain>=0.1.0
langchain-anthropic>=0.1.0
anthropic>=0.18.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Development dependencies
pytest>=8.0.0
pytest-cov>=4.1.0
pre-commit>=3.6.0
ruff>=0.8.0
pydocstyle>=6.3.0
bandit>=1.7.6
